pyproject.toml
[project]
name = "studentrentalsystem-python"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.3.7",
    "flask>=3.1.2",
    "ollama>=0.6.1",
    "pandas>=2.3.3",
    "pymongo>=4.15.5",
    "python-dotenv>=1.2.1",
    "redis>=7.1.0",
    "requests>=2.32.5",
    "selenium>=4.39.0",
    "streamlit>=1.52.1",
]
